+++ docs/tutorials/Tutorial_1-Document_Indexing/Step_1_CreateSyntheticData.sh
@@ -1,2 +1,2 @@
 #!/bin/bash
-chatterlang_script --script Step_1_CreateSyntheticData.script > stories.json
+chatterlang_script --script Step_1_CreateSyntheticData.script