+++ src/talkpipe/util/constants.py
@@ -1,0 +1,3 @@
+TALKPIPE_MODEL_NAME = "default_model_name"
+TALKPIPE_SOURCE = "default_model_source"
+OLLAMA_SERVER_URL = "OLLAMA_SERVER_URL"